pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "solfns"
version = "0.0.1-beta"
authors = [
  { name="Aaron Mishkin", email="amishkin@cs.stanford.edu" },
]
description = "Code for 'Optimal Sets and Solution Paths of ReLU Networks' presented at ICML 2023."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=1.21.3",
    "torch",
    "torchvision",
    "cvxpy>=1.2.1",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.2",
    "typing_extensions",
    "tqdm",
    "opt_einsum",
    "matplotlib",
    "linalg_backends",
    "pyscnn @ git+https://git@github.com/pilancilab/scnn@solfns#egg=scnn"
]

[tool.setuptools.package-data]
solfns = ["py.typed"]

[project.urls]
"Homepage" = "https://github.com/pilancilab/relu_optimal_sets"
"Bug Tracker" = "https://github.com/pilancilab/relu_optimal_sets/issues"
